src/requirements.txt
spotipy
openai
python-dotenv
flask
requests
requests-oauthlib